+++ src/main/java/com/gghyrmrwf/glebthanwolves/GlebThanWolves.java
@@ -2,12 +2,16 @@ package com.gghyrmrwf.glebthanwolves;
 
 import com.gghyrmrwf.glebthanwolves.events.AlwaysHostilePiglinsEvents;
 import com.gghyrmrwf.glebthanwolves.events.AlwaysHostileZombifiedPiglinsEvents;
+import com.gghyrmrwf.glebthanwolves.events.AngryHorseBuckEvents;
 import com.gghyrmrwf.glebthanwolves.events.BushcraftBreakEvents;
+import com.gghyrmrwf.glebthanwolves.events.ChickenLaysTntEvents;
 import com.gghyrmrwf.glebthanwolves.events.ExplodingPigEvents;
+import com.gghyrmrwf.glebthanwolves.events.ExplodingRabbitJumpEvents;
 import com.gghyrmrwf.glebthanwolves.events.GhastFollowRangeEvents;
 import com.gghyrmrwf.glebthanwolves.events.HardcoreEvents;
 import com.gghyrmrwf.glebthanwolves.events.MagmaBlockHazardEvents;
 import com.gghyrmrwf.glebthanwolves.events.NetherIdleHazardEvents;
+import com.gghyrmrwf.glebthanwolves.events.SheepShearedWitherEvents;
 import com.gghyrmrwf.glebthanwolves.events.WorldEvents;
 import com.mojang.logging.LogUtils;
 import net.minecraftforge.common.MinecraftForge;
@@ -43,6 +47,10 @@ public class GlebThanWolves {
         MinecraftForge.EVENT_BUS.register(new GhastFollowRangeEvents());
         MinecraftForge.EVENT_BUS.register(new MagmaBlockHazardEvents());
         MinecraftForge.EVENT_BUS.register(new ExplodingPigEvents());
+        MinecraftForge.EVENT_BUS.register(new ChickenLaysTntEvents());
+        MinecraftForge.EVENT_BUS.register(new SheepShearedWitherEvents());
+        MinecraftForge.EVENT_BUS.register(new ExplodingRabbitJumpEvents());
+        MinecraftForge.EVENT_BUS.register(new AngryHorseBuckEvents());
     }
 
     private void commonSetup(final FMLCommonSetupEvent event) {